requirements.txt
azure-functions
twilio==6.35.4